refactor(cli): move cpumask parsing to cli module
Parsing the hex string given for the `--cpumask` cli argument was
previously part of xnvmeperf, which would require us to duplicate or
rewrite this parser if the argument were to reused. I just move the
parser to the cli module, such that cli->args has the list of CPU ids in
the CPU field. Note that the `struct xnvme_cli_args` grows by 4KB

+++ tools/xnvmeperf/xnvmeperf.h
@@ -22,8 +22,8 @@ enum iopattern {
 struct xnvmeperf_args {
 	int ndevs;
 	const char **dev_uris;
-	int ncpus;
-	int *cpus;
+	uint16_t ncpus;
+	uint16_t *cpus;
 	uint32_t qdepth;
 	uint32_t iosize;
 	uint32_t time;
